requirements.txt
rawpy>=0.19.0
Pillow>=10.0.0
imagehash>=4.3.1
numpy>=1.24.0
tqdm>=4.66.0
matplotlib>=3.7.0
flask>=3.0.0
